+++ packages/poker-state-machine/state_machine.ts
@@ -1,11 +1,7 @@
 import { BehaviorSubject, Subject } from "rxjs";
-import {
-  determineHandType,
-  determineWinningHand,
-  getShuffledDeck,
-  type Card,
-  type Deck,
-} from "./poker";
+import type { Card, Deck } from "./poker";
+import { dealer, bigBlind, currentPlayer,playersInRound, roundRotation, seatedPlayers, smallBlind } from "./queries";
+import { processPlayerMove, startRound, transitionPhase } from "./transitions";
 
 // Indicates whether the player is playing the current round.
 // NOTE: FOLDED is just a temporary marker so the length of our
@@ -63,276 +59,9 @@ export type Move =
   | { type: "call" }
   | { type: "raise"; amount: number };
 
-// queries
-
-const players = (state: PokerState) => Object.values(state.players).sort()
-
-export const seatedPlayers = (state: PokerState) =>
-  Object.keys(state.players).length;
-
-export const playingPlayers = (state: PokerState) =>
-  players(state).filter((p) => p.status === "PLAYING");
-
-export const playersInRound = (state: PokerState) =>
-  players(state).filter(
-    p => p.status === "PLAYING" || p.status === "FOLDED",
-  );
-
-export const SMALL_BLIND = 10;
-export const BIG_BLIND = 20;
-
-export const dealer = (state: PokerState) =>
-  playersInRound(state)[state.dealerIndex];
-
-export function firstPlayerIndex(state: PokerState): number {
-  const players = playersInRound(state).length;
-  const preflop = state.community.length === 0;
-  return players === 2 && preflop
-    ? state.dealerIndex
-    : (state.dealerIndex + 3) % players;
-}
-
-export function rotated<T>(array: T[], count: number): T[] {
-  const length = array.length;
-  return Array.from({ length }).map((_, i) => array[(i + count) % length]);
-}
-
-export const roundRotation = (state: PokerState) => {
-  const players = playersInRound(state);
-  const index = firstPlayerIndex(state);
-  return rotated(players, index);
-};
-
-export const currentPlayer = (state: PokerState) => {
-  const players = roundRotation(state);
-  return players[state.currentPlayerIndex];
-};
-
-export const bigBlind = (state: PokerState) => {
-  const players = playersInRound(state);
-  return players[(state.dealerIndex + 1) % players.length];
-};
-
-export const smallBlind = (state: PokerState) => {
-  const players = playersInRound(state);
-  return players[(state.dealerIndex + 2) % players.length];
-};
-
-// state transition functions
-
-// precondition: waiting for players | finished previous round
-export function dealCards(state: PokerState): PokerState {
-  const deck = getShuffledDeck();
-  const dealtCards = deck.splice(0, 2 * seatedPlayers(state));
-
-  return {
-    ...state,
-    status: "PLAYING",
-    deck,
-    community: [],
-    burnt: [],
-    currentPlayerIndex: 0,
-    // TODO: use conjugateEntries
-    players: Object.fromEntries(
-      players(state).map((p, i) => [
-        p.id,
-        {
-          ...p,
-          hand: dealtCards.slice(2 * i, 2 * i + 2),
-          status: "PLAYING",
-          bet: 0,
-        },
-      ]),
-    ),
-  };
-}
-
-// precondition: waiting for players | finished previous round
-export function rotateBlinds(state: PokerState): PokerState {
-  const players = playersInRound(state).length;
-  return {
-    ...state,
-    dealerIndex: (state.dealerIndex + 1) % players,
-  };
-}
-
-// precondition: cards are dealt
-export function collectBlinds(state: PokerState): PokerState {
-  const bigBlindId = bigBlind(state).id;
-  const bigBlindPlayer = state.players[bigBlindId];
-  const bigBlindAmount = Math.min(BIG_BLIND, bigBlindPlayer.chips);
-
-  const smallBlindId = smallBlind(state).id;
-  const smallBlindPlayer = state.players[smallBlindId];
-  const smallBlindAmount = Math.min(SMALL_BLIND, smallBlindPlayer.chips);
-
-  return {
-    ...state,
-    pot: state.pot + bigBlindAmount + smallBlindAmount,
-    bet: Math.max(bigBlindAmount, smallBlindAmount),
-    players: {
-      ...state.players,
-      [bigBlindId]: {
-        ...bigBlindPlayer,
-        chips: bigBlindPlayer.chips - bigBlindAmount,
-        bet: bigBlindAmount,
-      },
-      [smallBlindId]: {
-        ...smallBlindPlayer,
-        chips: smallBlindPlayer.chips - smallBlindAmount,
-        bet: smallBlindAmount,
-      },
-    },
-  };
-}
-
-export function startRound(current: PokerState): PokerState {
-  const nextState = [dealCards, rotateBlinds, collectBlinds].reduce(
-    (state, f) => f(state),
-    current,
-  )
-
-  console.log('starting round', { nextState })
-  return nextState
-}
-
-function shiftBetRotation(state: PokerState): PokerState {
-  const players = roundRotation(state);
-  // needs to be >= because current player might have folded and be the last
-  const isLastPlayer = state.currentPlayerIndex >= players.findLastIndex((p) => p.status === "PLAYING");
-
-  const allCalled = players.every(
-    (p) => p.status === "FOLDED" || p.bet === state.bet,
-  );
-
-  const nextState = structuredClone(state)
-  const playersLeft = players.filter(p => p.status === "PLAYING");
-  if (playersLeft.length === 1) {
-    nextState.winningPlayerId = playersLeft[0].id;
-    nextState.currentPlayerIndex = -1
-    return nextState
-  }
-
-  if (isLastPlayer && allCalled) {
-    nextState.currentPlayerIndex = -1
-    return nextState
-  }
-
-  const nextPlayerIndex = isLastPlayer && !allCalled
-    ? players.findIndex((p) => p.status === "PLAYING") // rotates back to the beggining
-    : players.findIndex(
-      // find next player still in round
-      (p, i) => p.status === "PLAYING" && state.currentPlayerIndex < i,
-    );
-
-  nextState.currentPlayerIndex = nextPlayerIndex
-
-  return nextState;
-}
-
-function playerBet(state: PokerState, playerId: string, bet: number): PokerState {
-  const player = state.players[playerId]
-  const diff = Math.min(bet - player.bet, player.chips)
-  const amount = player.bet + diff
-  const remaining = player.chips - diff
-  const raised = bet > state.bet
-
-  return {
-    ...state,
-    pot: state.pot + diff,
-    bet: Math.max(state.bet, amount),
-    players: {
-      ...state.players,
-      [playerId]: {
-        ...player,
-        bet: amount,
-        chips: remaining,
-        status: 'PLAYING'
-      }
-    },
-  }
-}
-
-export function processPlayerMove(state: PokerState, move: Move): PokerState {
-  const players = roundRotation(state);
-  const playerId = players[state.currentPlayerIndex].id;
-
-  let nextState = structuredClone(state);
-  switch (move.type) {
-    case "fold": {
-      // TODO: check if the player already has enough bet
-      nextState.players[playerId].status = "FOLDED";
-      break;
-    }
-
-    case "call": {
-      nextState = playerBet(nextState, playerId, state.bet)
-      break;
-    }
-
-    case "raise": {
-      nextState = playerBet(nextState, playerId, move.amount)
-      break;
-    }
-  }
-
-  return shiftBetRotation(nextState);
-}
-
-// precondition: all players settled on a bet size
-export function transitionPhase(state: PokerState): PokerState {
-  if (state.winningPlayerId) {
-    return dealCards(state)
-  }
-
-  let nextState = structuredClone(state);
-  nextState.currentPlayerIndex = 0;
-  nextState.bet = 0
-  nextState.players = Object.fromEntries(Object.entries(state.players).map(([id, player]) => [id, { ...player, bet: 0 }]));
-
-  switch (state.community.length) {
-    // deal flop
-    case 0: {
-      // TODO: what to do about invalid states? (i.e. has burnt but no community)
-      nextState.burnt = [nextState.deck.pop()!];
-      nextState.community = [
-        nextState.deck.pop()!,
-        nextState.deck.pop()!,
-        nextState.deck.pop()!,
-      ];
-      break;
-    }
-    // deal turn
-    case 3: {
-      nextState.burnt.push(nextState.deck.pop()!);
-      nextState.community.push(nextState.deck.pop()!);
-      break;
-    }
-    // deal river
-    case 4: {
-      nextState.burnt.push(nextState.deck.pop()!);
-      nextState.community.push(nextState.deck.pop()!);
-      break;
-    }
-    // showdown
-    case 5: {
-      const players = playingPlayers(state);
-      const playerHands = players.map((p) => [...p.hand, ...state.community]);
-      const playerHandTypes = playerHands.map((h) => determineHandType(h));
-      const winningHand = determineWinningHand(playerHandTypes);
-
-      const winningPlayerIndex = playerHandTypes.findIndex(
-        (t) => t === winningHand,
-      );
-      const winningPlayer = players[winningPlayerIndex]
-
-      nextState.winningPlayerId = winningPlayer.id;
-      nextState.players[winningPlayer.id].chips += state.pot;
-    }
-  }
-
-  return nextState;
-}
+export type TableEvent =
+  | { type: 'join', playerId: string }
+  | { type: 'leave', playerId: string };
 
 export type PlayerView = {
   // here because sometimes agents don't bother reading it further inside
@@ -358,18 +87,25 @@ export class PokerRoomStateMachine {
   private movesSubject = new Subject<Move>();
   public moves$ = this.movesSubject.asObservable();
 
+  private tableEventsSubject = new Subject<TableEvent>();
+  public tableEvents$ = this.tableEventsSubject.asObservable();
+
   constructor(private minimumPlayers: number) {
     this.state$.subscribe((state) => {
+      // TODO: add a delay to start round
       if (state.status === "WAITING" && seatedPlayers(state) >= this.minimumPlayers) {
-        console.log('starting round...')
-        this.subject.next(startRound(state))
+        this.startRound();
         return;
       }
 
-      if (state.status === "PLAYING" && playersInRound(state).length === 1) {
-        // the remaining player wins
+      if (state.status === "PLAYING" && state.winningPlayerId) {
         console.log("winningPlayerId: ", state.winningPlayerId);
         return;
+      }
+
+      // TODO: handle error
+      if (state.status === 'PLAYING' && playersInRound(state).length === 1) {
+        throw 'inconsistent state round is over but there are no remaining players'
       }
 
       if (
@@ -378,6 +114,7 @@ export class PokerRoomStateMachine {
         state.currentPlayerIndex === -1
       ) {
         this.transitionPhase();
+        return;
       }
     });
   }
@@ -386,8 +123,10 @@ export class PokerRoomStateMachine {
     return this.subject.value
   }
 
+  // TODO: implement leave table as well
   addPlayer(id: string) {
     const state = this.subject.value;
+    this.tableEventsSubject.next({ type: 'join', playerId: id });
     this.subject.next({
       ...state,
       players: {
@@ -425,7 +164,12 @@ export class PokerRoomStateMachine {
 
   currentPlayerId() {
     const state = this.value;
-    return roundRotation(state)[state.currentPlayerIndex].id;
+    return currentPlayer(state)?.id!;
+  }
+
+  startRound() {
+    console.log('starting round...')
+    this.subject.next(startRound(this.value));
   }
 
   processPlayerMove(move: Move) {
